requirements.txt
asgiref==3.8.1
Django==5.0.6
django-debug-toolbar==4.4.2
django-extensions==3.2.3
djangorestframework==3.15.1
sqlparse==0.5.0
tzdata==2024.1
django-debug-toolbar 
django-debug-toolbar 
whitenoise==6.0.0
gitpython==3.1.31